refactor(navigation): add shared wheel navigation hook
